requirements.txt
pyautogui==0.9.54
opencv-python==4.8.1.78
Pillow==10.0.1
pygetwindow==0.0.9
